📦 NEW: add search functionality
