--- src/components/Search/index.tsx
@@ -1,7 +1,0 @@
-function Search() {
-  return (
-    <div>index</div>
-  )
-}
-
-export default Search
